[remove] #16 아티클 DTO 패키지 분리 (request/response)
