+++ src/main/java/org/farmsystem/sotserver/domain/article/dto/response/ArticleDetailResponse.java
@@ -1,10 +1,11 @@
-package org.farmsystem.sotserver.domain.article.dto;
+package org.farmsystem.sotserver.domain.article.dto.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.Builder;
+import lombok.Getter;
 import org.farmsystem.sotserver.domain.article.entity.Article;
 import org.farmsystem.sotserver.domain.article.entity.ArticleStatus;
+import org.farmsystem.sotserver.domain.comment.dto.CommentResponse;
 import org.farmsystem.sotserver.global.s3.S3Uploader;
 
 import java.time.LocalDateTime;
@@ -13,7 +14,7 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 @Builder
-public class ArticleListResponse {
+public class ArticleDetailResponse {
     private Long id;
     private String title;
     private String content;
@@ -23,23 +24,22 @@ public class ArticleListResponse {
     private LocalDateTime updateAt;
     private List<String> imageUrls;
 
-    // 댓글 개수 필드
-    private int commentCount;
+    // 댓글 정보 필드 추가
+    private List<CommentResponse> comments;
 
-    // 팩토리 메서드
-    public static ArticleListResponse from(Article article, S3Uploader s3Uploader, int commentCount) {
-        return ArticleListResponse.builder()
+    public static ArticleDetailResponse from(Article article, S3Uploader s3Uploader, List<CommentResponse> comments) {
+        return ArticleDetailResponse.builder()
                 .id(article.getArticleId())
                 .title(article.getTitle())
                 .content(article.getContent())
-                .status(article.getStatus())
                 .userId(article.getAuthor().getUserId())
+                .status(article.getStatus())
                 .createAt(article.getCreatedAt())
                 .updateAt(article.getUpdatedAt())
                 .imageUrls(article.getImages().stream()
                         .map(img -> s3Uploader.getFileUrl(img.getKey()))
                         .toList())
-                .commentCount(commentCount)
+                .comments(comments)
                 .build();
     }
 }
